requirements.txt
flask
flask_graphql
flask_cors
graphene
requests
simplejson
python3-memcached
# sudo apt-get install memcached
python3-memcached
websocket-client
python-memcached
sqlalchemy
pymysql
sqlalchemy-datatables
gevent